requirements.txt
Django
gunicorn
#psycopg2
psycopg2-binary
python-decouple